refactor: 어제 날짜 반환 함수 extension으로 전환
뷰컨에서도 쓰일거같음!
