+++ BoxOffice/Model/Extension.swift
@@ -26,3 +26,17 @@ extension BoxOfficeResult: Equatable {
         return true
     }
 }
+
+extension Date {
+    func yesterday() -> String {
+        let formatter = DateFormatter()
+        formatter.dateFormat = "yyyyMMdd"
+        
+        let yesterday = Calendar.current.date(byAdding: .day, value: -1, to: Date())
+        guard let yesterday = yesterday else {
+            return ""
+        }
+        
+        return formatter.string(from: yesterday)
+    }
+}
